requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==5.1.1
distro==1.9.0
executing==2.1.0
ffmpeg==1.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gmqtt==0.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.58.1
openai-whisper==20240930
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
propcache==0.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
websockets==14.1
wyoming==1.6.0
yarl==1.18.3
